logind: `lock` when lid is closed
`suspend` would leave the OS in an unresponsive state
when the lid was re-opened for 10+ seconds, after which
the entire system would reboot.

`lock` is also not great since the display doesn't seem
to turn off when the lid is closed, but I guess it's better
than rebooting every time.